Modify database tables
Add API hidden methods

+++ DessertTaCeinture.Database/DessertTaCeinture.Database/Tables/Recipe_Ingredients.sql
@@ -3,7 +3,7 @@
     [Id] INT NOT NULL IDENTITY(1,1), 
 	[RecipeId] INT NOT NULL, 
     [IngredientId] INT NOT NULL,  
-    [Quantity] FLOAT NOT NULL,  
+    [Quantity] INT NOT NULL,  
     [Unit] VARCHAR(50) NOT NULL,
     CONSTRAINT [PK_Recipe_Ingredients] PRIMARY KEY ([Id]), 
     CONSTRAINT [FK_Recipe_Ingredients_ToRecipe] FOREIGN KEY ([RecipeId]) REFERENCES [Recipe]([Id]),
